implemented delete function

+++ src/main/java/com/victor/demoapp/demoapitest/service/PersonService.java
@@ -14,6 +14,6 @@ public interface PersonService {
 //
 //	PersonResponse update(Long id, PersonRequest value);
 
-	void delete(Long id);
+	void delete(int id);
 	
 }
